requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.0.6
gunicorn==21.2.0
fastapi==0.115.7
flatbuffers==25.1.21
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
idna==3.10
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
namex==0.0.8
numpy==1.23.5
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==1.5.3
pillow==11.0.0
protobuf==3.20.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.2
tensorboard==2.16.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.17.2